Move daemon fixture up to conftest
